updated file structure for m2d1
